Address 2 fatal import issues.
Thanks to a recent improvement to loadDb() (see commit 98318 in AnnotationDbi),
which potentially affects any package that imports a TxDb package, doing
library(ChIPseeker) in a fresh R session now attaches only 1 package (the
ChIPseeker itself) instead of 11 before the change. Note that this is really
how it should be given that ChIPseeker doesn't depend on any package via its
Depends field. However, this change reveals many import issues with the
package. These issues are now detected and reported by 'R CMD check' (see all
the NOTEs about "no visible global function definition"). Before the change
to loadDb(), they weren't (see the build report for ChIPseeker in release).
I addressed only 2 of them (the fatal ones). Please take the time to address
the other ones when you get a chance. Let me know or ask on the bioc-devel
mailing list if you have questions or concerns about this. Thanks!  

On an unrelated note, I also replaced the use of elementMetadata() with mcols()
(they are the same but the latter is preferred over the former).


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/ChIPseeker@98566 bc3139a8-67e5-0310-9ffc-ced21a209358