Update plotAnno.R
This update is to fix issue#134 and it adds two functions to the plotAnno().
The first one is to add cex to legend() in order to adjust the size of labels, which can solve the problem of incomplete legends.
The second one is to add radius parameter in order to change the size of Pie. The default value of radius is 0.8. This parameter can be used when the legend.position != "rightside"(e.x "topright"). User can change the radius of the Pie by radius and change the size of labels by cex, which can meet the different needs of Users.
The last change is the default value of cex(from 0.9 to 0.8), which I think is more suitable for Users by my experience of testing painting.